+++ ext_tables.sql
@@ -1,4 +1,0 @@
-CREATE TABLE sys_file_processedfile (
-	`integration` VARCHAR(32),
-	`integration_checksum` VARCHAR(40)
-);